Boss Line Of Sight Implemented
The boss now wanders quickly around the dungeon, unless the player is
close to him in which case he slows down. If the player is close, and
within his view angle, he will attempt to see him. If there are no walls
in the way, he will chase the player.