Enemy Seek AI Improved
Enemies will now chase the player around a corner. If the player goes
around a corner, the AI will seek the players last known location, until
it either sees the player again or reaches the point. If the player is
not visible from the new point, they will return to wandering the
dungeon.